pyproject.toml
[project]
name = "alliance-fulfillment"
version = "0.1.0"
description = "Fulfillment queue dispenser for Alliance Chemical"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.0",
    "twilio>=9.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
]

[project.scripts]
fulfillment-web = "fulfillment.api:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
